# speed up `pivot_longer()`