--- a/R/pivot-long.R
+++ b/R/pivot-long.R
@@ -233,6 +233,8 @@ pivot_longer_spec <- function(data,
     val_type <- vec_ptype_common(!!!set_names(val_cols[col_id], cols), .ptype = values_ptypes[[value]])
     out <- vec_c(!!!val_cols, .ptype = val_type)
     # Interleave into correct order
+    # TODO somehow `t(matrix(x))` is _faster_ than `matrix(x, byrow = TRUE)`
+    # if this gets fixed in R this should use `byrow = TRUE` again
     n_vals <- nrow(data) * length(val_cols)
     idx <- t(matrix(seq_len(n_vals), ncol = n_vals / nrow(data)))
     vals[[value]] <- vec_slice(out, as.integer(idx))