Changes
* don't create `rows` data frame but slice the vectors directly
* don't use `matrix(byrow = TRUE)` as it is surprisingly slow

It looks like a bug to me that `matrix(x, byrow = TRUE)` is slower than `t(matrix(x))`. Probably, this should be fixed in base R.

## Benchmark: `pivot_longer()`

```r
library(data.table)
library(tidyr)

data.test <- matrix(
  data = sample(
    x = c(0L, 1L, 2L, NA_integer_),#the genotypes
    size = 2e+07,
    replace = TRUE,
    prob = c(0.8, 0.10, 0.05, 0.05)
  ),
  nrow = 20000,#number of SNPs/markers
  ncol = 1000,#number of samples
  dimnames = list(rownames = seq(1, 20000, 1), colnames = seq(1, 1000, 1))
) %>%
  tibble::as_tibble(x = ., rownames = "MARKERS")

bench::mark(
  pivot = tidyr::pivot_longer(data.test, cols = -MARKERS, names_to = "INDIVIDUALS", values_to = "GENOTYPES"),
  melt = data.table::as.data.table(data.test) %>%
    data.table::melt.data.table(
      data = .,
      id.vars = "MARKERS",
      variable.name = "INDIVIDUALS",
      value.name = "GENOTYPES",
      variable.factor = FALSE) %>%
    tibble::as_tibble(.),
  check = FALSE,
  iterations = 5
)

# This PR
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 2 x 13
#>   expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>
#> 1 pivot         452ms    465ms      1.96     918MB     5.89     5    15      2.54s <NULL>
#> 2 melt          247ms    362ms      2.65     459MB     2.65     5     5      1.88s <NULL>

# CRAN
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 2 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 pivot         1.27s    1.27s     0.790    1.12GB     3.16
#> 2 melt       264.05ms 338.71ms     2.95   458.57MB     1.48
```

## Benchmark: `matrix()`

``` r
library(ggplot2)
library(purrr)

matrix_compare <- function(n, ncol) {
  bench::mark(
    by_row = matrix(seq_len(n), ncol = ncol, byrow = TRUE),
    transpose = t(matrix(seq_len(n), ncol = n / ncol))
  )
}

n_bench <- map_dfr(
  set_names(10e3 * c(1L, 2L, 5L, 10L, 20L, 50L, 100L)),
  matrix_compare,
  ncol = 200,
  .id = "n"
)

n_bench %>% 
  dplyr::transmute(
    n = as.integer(n),
    name = names(expression),
    median
  ) %>% 
  ggplot(aes(n, median, colour = name)) +
  geom_line() +
  geom_point() +
  labs(
    colour = "approach",
    x = "sequence size",
    y = "median time"
  )
```

![](https://i.imgur.com/arPxWsz.png)

``` r
ncol_bench <- map_dfr(
  set_names(200L * c(1L, 2L, 5L, 10L, 20L, 50L, 100L)),
  matrix_compare,
  n = 1e6,
  .id = "ncol"
)

ncol_bench %>% 
  dplyr::transmute(
    ncol = as.integer(ncol),
    name = names(expression),
    median
  ) %>% 
  ggplot(aes(ncol, median, colour = name)) +
  geom_line() +
  geom_point() +
  labs(
    colour = "approach",
    x = "Number of columns",
    y = "median time"
  )
```

![](https://i.imgur.com/P2FCN6V.png)

<sup>Created on 2021-06-17 by the [reprex package](https://reprex.tidyverse.org) (v2.0.0)</sup>